stream command cannot be class of client
